Id Long타입 -> Int로 변경

+++ src/main/java/kr/ac/hansung/cse/repo/ProductRepository.java
@@ -1,17 +1,17 @@
 package kr.ac.hansung.cse.repo;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.data.repository.CrudRepository;
 
 import kr.ac.hansung.cse.model.Product;
 
 
-public interface ProductRepository extends CrudRepository<Product, Long>{ 
+public interface ProductRepository extends CrudRepository<Product, Integer>{ 
 	
 	List<Product> findAll();
 	
 	List<Product> findByCategory(String category);
 	
-
 }
